Update code structure and enhance agent conditions
In this commit, the C++ version has been updated to C++20 and the code structure has been edited to include more detailed interaction messages for bank depositing. The increase in Agent's fatigue rate has been amplified to fivefold. Additionally, two new conditions were added for the agent - to move to the stage of buying stuff or going home to sleep when fatigued. Furthermore, fixes for debugging were recorded in Instructions file.