Refactor workplace assignment in AgentOwnedStates
This commit includes refactoring of the method used to assign workplaces to agents in the AgentOwnedStates class. The agent's specific location is now checked within the nested if statements, to ensure that agents are only assigned to workplaces that they are not currently located at. This improves the randomness and diversity of the selected workplaces.